chore: page 디렉토리 구분
