+++ src/pages/auth/LoginPage.tsx
@@ -1,15 +1,15 @@
-import Input from '../components/input/text/Input';
-import Button from '../components/input/button/Button';
-import Divider from '../components/divider/Divider';
+import Input from '../../components/input/text/Input';
+import Button from '../../components/input/button/Button';
+import Divider from '../../components/divider/Divider';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
 import React, { useContext, useEffect, useState } from 'react';
-import { LoginRequest } from '../types/auth';
-import useLogin from '../hooks/auth/useLogin';
-import useInvalid from '../hooks/valid/useInvalid';
-import { ToastContext } from '../contexts/ToastContext';
+import { LoginRequest } from '../../types/auth';
+import useLogin from '../../hooks/auth/useLogin';
+import useInvalid from '../../hooks/valid/useInvalid';
+import { ToastContext } from '../../contexts/ToastContext';
 import { AxiosError } from 'axios';
-import { ErrorResponse } from '../types/axios';
-import { AuthenticationContext } from '../contexts/AuthenticationContext';
+import { ErrorResponse } from '../../types/axios';
+import { AuthenticationContext } from '../../contexts/AuthenticationContext';
 
 const LoginPage = () => {
   const navigate = useNavigate();
