+++ src/pages/question/ExplainPage.tsx
@@ -1,6 +1,6 @@
-import Navigation from '../components/navigation/Navigation';
-import BackButton from '../components/navigation/BackButton';
-import QuillViewer from '../components/editor/QuillViewer';
+import Navigation from '../../components/navigation/Navigation';
+import BackButton from '../../components/navigation/BackButton';
+import QuillViewer from '../../components/editor/QuillViewer';
 import { useLocation } from 'react-router-dom';
 
 const ExplainPage = () => {
