+++ src/pages/question/QuestionSubmitPage.tsx
@@ -1,18 +1,18 @@
-import Navigation from '../components/navigation/Navigation';
-import BackButton from '../components/navigation/BackButton';
-import MultiLineInput from '../components/input/text/MultiLineInput';
-import Button from '../components/input/button/Button';
-import Icon from '../components/icon/Icon';
+import Navigation from '../../components/navigation/Navigation';
+import BackButton from '../../components/navigation/BackButton';
+import MultiLineInput from '../../components/input/text/MultiLineInput';
+import Button from '../../components/input/button/Button';
+import Icon from '../../components/icon/Icon';
 import { faAngleLeft, faAngleRight } from '@fortawesome/free-solid-svg-icons';
 import { useNavigate, useParams } from 'react-router-dom';
-import useSubmitQuestion from '../hooks/question/useSubmitQuestion';
-import useFetchQuestion from '../hooks/question/useFetchQuestion';
+import useSubmitQuestion from '../../hooks/question/useSubmitQuestion';
+import useFetchQuestion from '../../hooks/question/useFetchQuestion';
 import { useState } from 'react';
-import { QuestionSubmitRequest } from '../types/question';
-import Modal from '../components/modal/Modal';
-import OutlineButton from '../components/input/button/OutlineButton';
-import useInvalid from '../hooks/valid/useInvalid';
-import useQuestionNavigation from '../hooks/question/useQuestionNavigation';
+import { QuestionSubmitRequest } from '../../types/question';
+import Modal from '../../components/modal/Modal';
+import OutlineButton from '../../components/input/button/OutlineButton';
+import useInvalid from '../../hooks/valid/useInvalid';
+import useQuestionNavigation from '../../hooks/question/useQuestionNavigation';
 
 const QuestionSubmitPage = () => {
   const navigate = useNavigate();
